chore: update exports
