+++ packages/network/index.ts
@@ -1,8 +1,5 @@
-export * as aptosHandler from './src/handlers/aptos';
-export * as solanaHandler from './src/handlers/solana';
-export * as suiHandler from './src/handlers/sui';
-export * as tezosHandler from './src/handlers/tezos';
+export * as aptos from './src/aptos';
 export * as solana from './src/solana';
-export * from './src/sui';
-export * from './src/tezos';
+export * as sui from './src/sui';
+export * as tezos from './src/tezos';
 export * as utils from './src/utils';
