+++ packages/network/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@walless/network",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "main": "./index.ts",
   "types": "./index.ts",
   "license": "MIT",
