--- packages/network/src/utils/connection.ts
@@ -1,3 +1,0 @@
-import { clusterApiUrl, Connection } from '@solana/web3.js';
-
-export const connection = new Connection(clusterApiUrl('devnet'));
